requirements.txt
annotated-types==0.7.0
blobfile==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
einops==0.8.1
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
MedPy==0.5.2
monai==1.3.2
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
opencv-python==4.12.0.88
packaging==25.0
pillow==10.2.0
platformdirs==4.3.6
prettytable==3.11.0
protobuf==5.29.5
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==2.34.1
SimpleITK==2.5.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tifffile==2023.7.10
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.13.2
urllib3==2.2.3
wandb==0.21.0
wcwidth==0.2.13
zipp==3.20.2
